requirements.txt
python-dateutil
google-generativeai
rich
